[Chord] Degree is no longer calculated for all nodes
The degree wasn't used anywhere so it was removed from the base queries to speed up the process.
The degree was used only when generating sample visualization in which case the node degrees have to be explicitly requested.
Also the API has been standardize as it now always returns Node objects and not just uris.