[Core] Check connection to Virtuoso when platform SPA loads
Unavailable Virtuoso leads sometimes to weird behavior. To avoid those confusing situations, we make a simple SPARQL query to Virtuoso whenever user opens the platform SPA. This will probably results in a lots of redundant polling but given how the generator will be used (if at all) in the near future, we want to give the smoothest possible experience to anyone who decides to experiment with this tool.