[Core] Fixed ugly bug in promise duck API
createPromiseStatusesSelector queries all statuses of given name. If there is none, it just returns an empty map. That is dangerous in aggregation
(which happened for Google Maps). I'm waiting for several promises to finish and if some of them haven't even started, i.e., they don't appear in the state,
but the rest of them are finished, then the aggregated promise status is marked as done (even though there are promises that haven't even finish).

createPromiseStatusesSelector now returns always at least one promise status (by default the empty one) which makes the aggregation work again.